[TASK] Move generateMatchExcludes to utils
